Melhorias no Código para diminuir o número de arquivos e deixar mais fácil de compreender.
